scada_master/requirements.txt
# SCADA Master Requirements

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# WebSocket
websockets==12.0
aiohttp==3.9.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
pyjwt==2.8.0

# Database
asyncpg==0.29.0
aioredis==2.0.1

# Data Validation
pydantic==2.5.0

# Utilities
python-dateutil==2.8.2
